Do not call extract_version_to_git.delay anymore
