+++ src/olympia/lib/settings_base.py
@@ -1177,7 +1177,6 @@ CELERY_TASK_ROUTES = {
     'olympia.files.tasks.extract_webext_permissions': {'queue': 'addons'},
     'olympia.files.tasks.hide_disabled_files': {'queue': 'addons'},
     'olympia.versions.tasks.delete_preview_files': {'queue': 'addons'},
-    'olympia.versions.tasks.extract_version_to_git': {'queue': 'addons'},
     'olympia.git.tasks.continue_git_extraction': {'queue': 'addons'},
     'olympia.git.tasks.extract_versions_to_git': {'queue': 'addons'},
     'olympia.git.tasks.on_extraction_error': {'queue': 'addons'},
